+++ cmd/varlogadm/app/cli.go
@@ -1,138 +1,96 @@
 package app
 
 import (
-	"fmt"
-	"net"
+	"path/filepath"
 
 	"github.com/urfave/cli/v2"
+	"go.uber.org/zap"
 
 	"github.daumkakao.com/varlog/varlog/internal/varlogadm"
 	"github.daumkakao.com/varlog/varlog/pkg/types"
+	"github.daumkakao.com/varlog/varlog/pkg/util/log"
 )
 
-func InitCLI(options *varlogadm.Options) *cli.App {
+func InitCLI() *cli.App {
 	app := &cli.App{
 		Name:    "varlogadm",
-		Usage:   "run varlog manager server",
+		Usage:   "run varlog admin server",
 		Version: "v0.0.1",
 	}
 	cli.VersionFlag = &cli.BoolFlag{Name: "version"}
-	app.Commands = append(app.Commands, initStartCommand(options))
+	app.Commands = append(app.Commands, initStartCommand())
 	return app
 }
 
-func initStartCommand(options *varlogadm.Options) *cli.Command {
-	startCmd := &cli.Command{
+func initStartCommand() *cli.Command {
+	return &cli.Command{
 		Name:    "start",
 		Aliases: []string{"s"},
 		Usage:   "start [flags]",
-		Action: func(c *cli.Context) error {
-			// ClusterID
-			parsedClusterID := c.Uint("cluster-id")
-			clusterID, err := types.NewClusterIDFromUint(parsedClusterID)
-			if err != nil {
-				return err
-			}
-			options.ClusterID = clusterID
-
-			// Metadata Repository Addresses
-			parsedMRAddrs := c.StringSlice("mr-address")
-			for _, addr := range parsedMRAddrs {
-				host, _, err := net.SplitHostPort(addr)
-				if err != nil {
-					return fmt.Errorf("invalid metadata repository address %v: %v", addr, err)
-				}
-				if ip := net.ParseIP(host); ip == nil {
-					return fmt.Errorf("invalid metadata repository address: %v", addr)
-				}
-			}
-			options.MetadataRepositoryAddresses = parsedMRAddrs
-
-			return Main(options)
+		Action:  start,
+		Flags: []cli.Flag{
+			flagClusterID.StringFlag(false, types.ClusterID(1).String()),
+			flagListen.StringFlag(false, varlogadm.DefaultListenAddress),
+			flagReplicationFactor.UintFlag(false, varlogadm.DefaultReplicationFactor),
+			flagMetadataRepository.StringSliceFlag(true, nil),
+			flagInitMRConnRetryCount.IntFlag(false, varlogadm.DefaultInitialMRConnectRetryCount),
+			flagInitMRConnRetryBackoff.DurationFlag(false, varlogadm.DefaultInitialMRConnectRetryBackoff),
+			flagSNWatcherRPCTimeout.DurationFlag(false, varlogadm.DefaultWatcherRPCTimeout),
+			flagMRConnTimeout.DurationFlag(false, varlogadm.DefaultMRConnTimeout),
+			flagMRCallTimeout.DurationFlag(false, varlogadm.DefaultMRCallTimeout),
+			flagLogDir.StringFlag(false, ""),
+			flagLogToStderr.BoolFlag(),
 		},
 	}
-	startCmd.Flags = []cli.Flag{
-		&cli.BoolFlag{
-			Name:        "verbose",
-			Aliases:     []string{"v"},
-			Value:       false,
-			Usage:       "verbose output",
-			EnvVars:     []string{"VERBOSE"},
-			Destination: &options.Verbose,
-		},
-	}
-	startCmd.Flags = append(startCmd.Flags,
-		&cli.UintFlag{
-			Name:    "cluster-id",
-			Aliases: []string{},
-			Value:   uint(varlogadm.DefaultClusterID),
-			Usage:   "cluster id",
-			EnvVars: []string{"CLUSTER_ID"},
-		},
-		&cli.StringFlag{
-			Name:        "listen-address",
-			Aliases:     []string{"rpc-bind-address"},
-			Value:       varlogadm.DefaultListenAddress,
-			Usage:       "RPC listemn address",
-			EnvVars:     []string{"LISTEN_ADDRESS", "RPC_BIND_ADDRESS"},
-			Destination: &options.ListenAddress,
-		},
-		&cli.UintFlag{
-			Name:        "replication-factor",
-			Aliases:     []string{},
-			Value:       varlogadm.DefaultReplicationFactor,
-			Usage:       "replication factor",
-			EnvVars:     []string{"REPLICATION_FACTOR"},
-			Destination: &options.ReplicationFactor,
-		},
-		&cli.StringSliceFlag{
-			Name:        "mr-address",
-			Aliases:     []string{},
-			Usage:       "metadata repository address",
-			EnvVars:     []string{"MR_ADDRESS"},
-			Destination: cli.NewStringSlice(),
-		},
-		&cli.IntFlag{
-			Name:        "init-mr-conn-retry-count",
-			Aliases:     []string{},
-			Value:       varlogadm.DefaultInitialMRConnectRetryCount,
-			Usage:       "the number of retry of initial metadata repository connect",
-			EnvVars:     []string{"INIT_MR_CONN_RETRY_COUNT"},
-			Destination: &options.InitialMRConnRetryCount,
-		},
-		&cli.DurationFlag{
-			Name:        "init-mr-conn-retry-backoff",
-			Aliases:     []string{},
-			Value:       varlogadm.DefaultInitialMRConnectRetryBackoff,
-			Usage:       "backoff duration between retries of initial metadata repository connect",
-			EnvVars:     []string{"INIT_MR_CONN_RETRY_BACKOFF"},
-			Destination: &options.InitialMRConnRetryBackoff,
-		},
-		&cli.DurationFlag{
-			Name:        "sn-watcher-rpc-timeout",
-			Aliases:     []string{},
-			Value:       varlogadm.DefaultWatcherRPCTimeout,
-			Usage:       "sn watcher rpc timeout",
-			EnvVars:     []string{"SN_WATCHER_RPC_TIMEOUT"},
-			Destination: &options.WatcherOptions.RPCTimeout,
-		},
-		&cli.DurationFlag{
-			Name:        "mr-conn-timeout",
-			Aliases:     []string{},
-			Value:       varlogadm.DefaultMRConnTimeout,
-			Usage:       "mr connection timeout",
-			EnvVars:     []string{"MR_CONN_TIMEOUT"},
-			Destination: &options.WatcherOptions.RPCTimeout,
-		},
-		&cli.DurationFlag{
-			Name:        "mr-call-timeout",
-			Aliases:     []string{},
-			Value:       varlogadm.DefaultMRCallTimeout,
-			Usage:       "mr call timeout",
-			EnvVars:     []string{"MR_CALL_TIMEOUT"},
-			Destination: &options.WatcherOptions.RPCTimeout,
-		},
-	)
+}
 
-	return startCmd
+func start(c *cli.Context) error {
+	clusterID, err := types.ParseClusterID(c.String(flagClusterID.Name))
+	if err != nil {
+		return err
+	}
+	logger, err := newLogger(c)
+	if err != nil {
+		return err
+	}
+	logger = logger.Named("adm").With(zap.Uint32("cid", uint32(clusterID)))
+	defer func() {
+		_ = logger.Sync()
+	}()
+
+	opts := varlogadm.DefaultOptions()
+	opts.Logger = logger
+	opts.MetadataRepositoryAddresses = c.StringSlice(flagMetadataRepository.Name)
+	opts.ListenAddress = c.String(flagListen.Name)
+	opts.ReplicationFactor = c.Uint(flagReplicationFactor.Name)
+	opts.InitialMRConnRetryCount = c.Int(flagInitMRConnRetryCount.Name)
+	opts.InitialMRConnRetryBackoff = c.Duration(flagInitMRConnRetryBackoff.Name)
+	opts.WatcherOptions.RPCTimeout = c.Duration(flagSNWatcherRPCTimeout.Name)
+	opts.MRManagerOptions.ConnTimeout = c.Duration(flagMRConnTimeout.Name)
+	opts.MRManagerOptions.CallTimeout = c.Duration(flagMRCallTimeout.Name)
+	return Main(opts)
+}
+
+func newLogger(c *cli.Context) (*zap.Logger, error) {
+	logtostderr := c.Bool(flagLogToStderr.Name)
+	logdir := c.String(flagLogDir.Name)
+	if !logtostderr && len(logdir) == 0 {
+		return zap.NewNop(), nil
+	}
+
+	logOpts := []log.Option{
+		log.WithHumanFriendly(),
+		log.WithZapLoggerOptions(zap.AddStacktrace(zap.DPanicLevel)),
+	}
+	if !logtostderr {
+		logOpts = append(logOpts, log.WithoutLogToStderr())
+	}
+	if len(logdir) != 0 {
+		absDir, err := filepath.Abs(logdir)
+		if err != nil {
+			return nil, err
+		}
+		logOpts = append(logOpts, log.WithPath(filepath.Join(absDir, "varlogadm.log")))
+	}
+	return log.New(logOpts...)
 }