+++ cmd/varlogadm/main.go
@@ -7,12 +7,10 @@ import (
 	_ "go.uber.org/automaxprocs"
 
 	"github.daumkakao.com/varlog/varlog/cmd/varlogadm/app"
-	"github.daumkakao.com/varlog/varlog/internal/varlogadm"
 )
 
 func main() {
-	options := varlogadm.DefaultOptions()
-	app := app.InitCLI(&options)
+	app := app.InitCLI()
 	if err := app.Run(os.Args); err != nil {
 		log.Printf("varlogadm: %v", err)
 		os.Exit(1)