+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Added `--server-max-msg-size` option to storage node. (#VARLOG-688/#601)
+- Added `--logdir` and `--logtostderr` options to varlogadm. (#VARLOG-692/#604)
 
 ### Changed
 - Changed default storage settings for good performance in usual cases. (#VARLOG-668/#584)